requirements.txt
coqui-tts[codec]
transformers>=4.57,<5.0
silero-vad
sounddevice
numpy
huggingface-hub
